src/bee_health_model/requirements.txt
numpy
pandas
tensorflow
fastapi
Pillow
uvicorn
python-multipart